Changes to library building
