+++ code/lib/SCC2691/src/SCC2691.c
@@ -1,6 +1,16 @@
 #include "SCC2691.h"
 #include "z80core.h"
 
+struct s_port
+{
+	unsigned char id;
+	unsigned char addr;
+	
+	unsigned char configReg[REG_CTLR];
+	unsigned char enabled;
+};
+struct s_port ports[SCC2691_PORTS];
+
 //=================================================================================
 // PRIVATE FUNCTIONS:
 //=================================================================================
@@ -50,22 +60,4 @@ int SCC2691_isEnabled(unsigned int port)
 {
 	if (ports[port].enabled)	return 1;
 	else						return 0;
-}
-
-int SCC2691_write(unsigned int port, char data)
-{
-	if (!ports[port].enabled)	return -1;
-	
-	_write_port(GET_REG_ADDR(port, REG_THR), data);
-	
-	return 0;
-}
-
-int SCC2691_read(unsigned int port, char* data)
-{	
-	if (!ports[port].enabled)	return -1;
-	
-	*data = _read_port(GET_REG_ADDR(port, REG_RHR));
-	
-	return 0;
 }
