+++ code/test/printf/printf.mk
@@ -1,5 +1,5 @@
 AREA      :=
 MAINFILE  := main.c
-LIBRARIES := z80sdcc SCC2691 z80core
+LIBRARIES := z80std SCC2691 z80core
 OUTPUT    := printf_test.bin
 PAD_BYTE  := 55
